+++ src/game/ability_layers.rs
@@ -291,6 +291,7 @@ impl Game {
             | AppliedEffectDef::CannotChangeController
             | AppliedEffectDef::RemainsAttachedThroughProtection
             | AppliedEffectDef::CannotBeBlockedBy(_)
+            | AppliedEffectDef::CanBlockOnly(_)
             | AppliedEffectDef::PreventDamageFrom(_)
             | AppliedEffectDef::PreventCombatDamage
             | AppliedEffectDef::AddLandTypes(_)