+++ src/game/continuous_effects.rs
@@ -237,6 +237,7 @@ impl Game {
             | AppliedEffectDef::CannotChangeController
             | AppliedEffectDef::RemainsAttachedThroughProtection
             | AppliedEffectDef::CannotBeBlockedBy(_)
+            | AppliedEffectDef::CanBlockOnly(_)
             | AppliedEffectDef::PreventDamageFrom(_)
             | AppliedEffectDef::PreventCombatDamage
             | AppliedEffectDef::AddLandTypes(_)