+++ src/game/entry_replacements.rs
@@ -569,6 +569,7 @@ impl Game {
             | AppliedEffectDef::CannotChangeController
             | AppliedEffectDef::RemainsAttachedThroughProtection
             | AppliedEffectDef::CannotBeBlockedBy(_)
+            | AppliedEffectDef::CanBlockOnly(_)
             | AppliedEffectDef::PreventDamageFrom(_)
             | AppliedEffectDef::PreventCombatDamage
             | AppliedEffectDef::AddLandTypes(_)